+++ dsort/request_spec.go
@@ -91,8 +91,8 @@ type RequestSpec struct {
 	MaxMemUsage       string        `json:"max_mem_usage"`             // Default: "80%"
 	BckProvider       string        `json:"bprovider"`                 // Default: "local"
 	OutputBckProvider string        `json:"output_bprovider"`          // Default: "local"
-	ExtractConcLimit  int64         `json:"extract_concurrency_limit"` // Default: DefaultConcLimit
-	CreateConcLimit   int64         `json:"create_concurrency_limit"`  // Default: DefaultConcLimit
+	ExtractConcLimit  int           `json:"extract_concurrency_limit"` // Default: DefaultConcLimit
+	CreateConcLimit   int           `json:"create_concurrency_limit"`  // Default: DefaultConcLimit
 	StreamMultiplier  int           `json:"stream_multiplier"`         // Default: transport.IntraBundleMultiplier
 	ExtendedMetrics   bool          `json:"extended_metrics"`          // Default: false
 	DryRun            bool          `json:"dry_run"`                   // Default: false
@@ -113,8 +113,8 @@ type ParsedRequestSpec struct {
 	OrderFileSep      string                `json:"order_file_sep"`
 	MaxMemUsage       cmn.ParsedQuantity    `json:"max_mem_usage"`
 	TargetOrderSalt   []byte                `json:"target_order_salt"`
-	ExtractConcLimit  int64                 `json:"extract_concurrency_limit"` // TODO: should be removed
-	CreateConcLimit   int64                 `json:"create_concurrency_limit"`  // TODO: should be removed
+	ExtractConcLimit  int                   `json:"extract_concurrency_limit"` // TODO: should be removed
+	CreateConcLimit   int                   `json:"create_concurrency_limit"`  // TODO: should be removed
 	StreamMultiplier  int                   `json:"stream_multiplier"`         // TODO: should be removed
 	ExtendedMetrics   bool                  `json:"extended_metrics"`
 	DryRun            bool                  `json:"dry_run"`
